// client/stocklinx app/src/redux/model/constant.ts
export enum modelConst {
  FETCH_MODELS_REQUEST = "FETCH_MODELS_REQUEST",
  FETCH_MODELS_SUCCESS = "FETCH_MODELS_SUCCESS",
  FETCH_MODELS_FAILURE = "FETCH_MODELS_FAILURE",
  FETCH_MODEL_REQUEST = "FETCH_MODEL_REQUEST",
  FETCH_MODEL_SUCCESS = "FETCH_MODEL_SUCCESS",
  FETCH_MODEL_FAILURE = "FETCH_MODEL_FAILURE",
  FETCH_MODEL_COUNTS_REQUEST = "FETCH_MODEL_COUNTS_REQUEST",
  FETCH_MODEL_COUNTS_SUCCESS = "FETCH_MODEL_COUNTS_SUCCESS",
  FETCH_MODEL_COUNTS_FAILURE = "FETCH_MODEL_COUNTS_FAILURE",
  CREATE_MODEL_REQUEST = "CREATE_MODEL_REQUEST",
  CREATE_MODEL_SUCCESS = "CREATE_MODEL_SUCCESS",
  CREATE_MODEL_FAILURE = "CREATE_MODEL_FAILURE",
  CREATE_RANGE_MODEL_REQUEST = "CREATE_RANGE_MODEL_REQUEST",
  CREATE_RANGE_MODEL_SUCCESS = "CREATE_RANGE_MODEL_SUCCESS",
  CREATE_RANGE_MODEL_FAILURE = "CREATE_RANGE_MODEL_FAILURE",
  UPDATE_MODEL_REQUEST = "UPDATE_MODEL_REQUEST",
  UPDATE_MODEL_SUCCESS = "UPDATE_MODEL_SUCCESS",
  UPDATE_MODEL_FAILURE = "UPDATE_MODEL_FAILURE",
  REMOVE_MODEL_REQUEST = "REMOVE_MODEL_REQUEST",
  REMOVE_MODEL_SUCCESS = "REMOVE_MODEL_SUCCESS",
  REMOVE_MODEL_FAILURE = "REMOVE_MODEL_FAILURE",
  REMOVE_RANGE_MODEL_REQUEST = "REMOVE_RANGE_MODEL_REQUEST",
  REMOVE_RANGE_MODEL_SUCCESS = "REMOVE_RANGE_MODEL_SUCCESS",
  REMOVE_RANGE_MODEL_FAILURE = "REMOVE_RANGE_MODEL_FAILURE",
  FILTER_MODELS_REQUEST = "FILTER_MODELS_REQUEST",
  FILTER_MODELS_SUCCESS = "FILTER_MODELS_SUCCESS",
  FILTER_MODELS_FAILURE = "FILTER_MODELS_FAILURE",
  SET_MODEL = "SET_MODEL",
  CLEAR_MODEL = "CLEAR_MODEL",
  SET_MODELS = "SET_MODELS",
  CLEAR_MODELS = "CLEAR_MODELS",
}